% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.predict.R
\name{as.data.frame.predictNestedLogit}
\alias{as.data.frame.predictNestedLogit}
\title{Convert a predictNestedLogit object to a data.frame}
\usage{
\method{as.data.frame}{predictNestedLogit}(x, row.names = NULL, optional = FALSE, newdata, ...)
}
\arguments{
\item{x}{a \code{predictNestedLogit} object}

\item{row.names}{row.names for result (for conformity with generic; not currently used)}

\item{optional}{logical. If TRUE, setting row names and converting column names
(to syntactic names: see make.names) is optional}

\item{newdata}{A  \code{newdata} data.frame used to generate predicted values. If not supplied,
the original data frame is used.}

\item{...}{other arguments (unused)}
}
\value{
A data frame containing the newdata values of predictors along with the columns
\code{response}, \code{p}, \code{se.p}, \code{logit}, \code{se.logit}
}
\description{
Convert a predictNestedLogit object to a data.frame
}
\examples{
data("Womenlf", package="carData")
comparisons <- logits(work=dichotomy("not.work", c("parttime", "fulltime")),
                     full=dichotomy("parttime", "fulltime"))

wlf.nested <- nestedLogit(partic ~ hincome + children,
                          dichotomies = comparisons,
                          data=Womenlf)
# get predicted values for a grid of `hincome` and `children`
new <- expand.grid(hincome=seq(0, 45, length=10),
                   children=c("absent", "present"))

pred.nested <- predict(wlf.nested, new)
plotdata <- as.data.frame(pred.nested, newdata=new)
str(plotdata)
}
