\name{GSS}
\alias{GSS}
\docType{data}
\title{
General Social Survey: Parent and Child Education
}
\description{
Comes from the General Social Survey ...
}
\format{
  A data.frame of 44091 observations, with the following variables
  \describe{
    \item{\code{parentdegree}}{parent educational degree, a factor with levels \code{l.t.highschool}
    \code{highschool} \code{college} \code{graduate}}
    \item{\code{degree}}{parent educational degree, a factor with levels \code{l.t.highschool}
    \code{highschool} \code{college} \code{graduate}}
    \item{\code{year}}{numeric year, 1972 - 2016}
    }
  }
\details{
No details yet
}
\source{
No source yet
}
\references{
No references yet
}
\examples{
summary(GSS)
}
\keyword{datasets}
