% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedMethods.R
\name{print.nested}
\alias{print.nested}
\alias{nestedMethods}
\alias{summary.nested}
\alias{print.summary.nested}
\alias{Anova.nested}
\alias{print.Anova.nested}
\alias{update.nested}
\alias{predict.nested}
\alias{coef.nested}
\alias{vcov.nested}
\alias{print.dichotomies}
\alias{as.dichotomies.matrix}
\alias{as.matrix.continuationDichotomies}
\alias{as.character.dichotomies}
\alias{as.matrix.dichotomies}
\alias{anova.nested}
\alias{print.anova.nested}
\alias{as.dichotomies}
\alias{glance.nested}
\alias{tidy.nested}
\title{Methods for \code{"nested"} and Related Objects}
\usage{
\method{print}{nested}(x, ...)

\method{summary}{nested}(object, ...)

\method{print}{summary.nested}(x, ...)

\method{Anova}{nested}(mod, ...)

\method{print}{Anova.nested}(x, ...)

\method{print}{dichotomies}(x, ...)

\method{predict}{nested}(object, newdata, ...)

\method{coef}{nested}(object, as.matrix = TRUE, ...)

\method{vcov}{nested}(object, as.matrix = FALSE, ...)

\method{anova}{nested}(object, object2, ...)

\method{print}{anova.nested}(x, ...)

\method{update}{nested}(object, formula, dichotomies, data, subset, contrasts, ...)

\method{as.matrix}{dichotomies}(x, ...)

\method{as.character}{dichotomies}(x, ...)

\method{as.matrix}{continuationDichotomies}(x, ...)

as.dichotomies(x, ...)

\method{as.dichotomies}{matrix}(x, ...)

\method{glance}{nested}(x, ...)

\method{tidy}{nested}(x, ...)
}
\arguments{
\item{x, object, object2, mod}{in most cases, an object of class \code{"nested"}.}

\item{\dots}{arguments to be passed down.}

\item{newdata}{a data frame containing combinations of values of the predictors
at which fitted probabilities are to be computed.}

\item{as.matrix}{if \code{TRUE} (the default for \code{coef}) return coefficients
as a matrix with one column for each nested dichotomy,
or coefficient covariances as a matrix with one row and column for each
combination of dichotomies and coefficients; if \code{FALSE} (the default for
\code{vcov}), return a list of coefficients or coefficient covariances
with one element for each dichotomy.}

\item{formula}{optional updated model formula.}

\item{dichotomies}{optional updated dichotomies object.}

\item{data}{optional updated data argument}

\item{subset}{optional updated subset argument.}

\item{contrasts}{optional updated contrasts argument.}
}
\description{
Various methods for processing \code{"nested"} and related objects.
Most of these are the standard methods for a model-fitting function.

\describe{
\item{\code{coef}, \code{vcov}}{return the coefficients and their variance-covariance matrix respectively.}
\item{\code{Anova}}{calculates type-II or type-III analysis-of-variance tables for \code{"nested"} objects.}
\item{\code{anova}}{computes analysis of variance (or deviance) tables for one or more fitted \code{"nested"}  objects.}
\item{\code{update}}{will re-fit a \code{"nested"} model with a change in any of the \code{formula}, \code{dichotomies},
\code{data}, \code{subset}, \code{contrasts}, ...}
}
}
\examples{

# TODO: add examples for at least some of the methods
}
\seealso{
\code{\link{nestedLogit}}
}
\author{
John Fox and Michael Friendly
}
\keyword{regression}
