% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedMethods.R
\name{nestedMethods}
\alias{nestedMethods}
\alias{print.nestedLogit}
\alias{summary.nestedLogit}
\alias{print.summary.nestedLogit}
\alias{Anova.nestedLogit}
\alias{print.Anova.nestedLogit}
\alias{update.nestedLogit}
\alias{predict.nestedLogit}
\alias{coef.nestedLogit}
\alias{vcov.nestedLogit}
\alias{print.dichotomies}
\alias{as.dichotomies.matrix}
\alias{as.matrix.continuationDichotomies}
\alias{as.character.dichotomies}
\alias{as.matrix.dichotomies}
\alias{linearHypothesis.nestedLogit}
\alias{print.predictDichotomies}
\alias{anova.nestedLogit}
\alias{print.anova.nestedLogit}
\alias{as.dichotomies}
\alias{glance.nestedLogit}
\alias{tidy.nestedLogit}
\title{Methods for \code{"nestedLogit"} and Related Objects}
\usage{
\method{print}{nestedLogit}(x, ...)

\method{summary}{nestedLogit}(object, ...)

\method{print}{summary.nestedLogit}(x, ...)

\method{Anova}{nestedLogit}(mod, ...)

\method{print}{Anova.nestedLogit}(x, ...)

\method{linearHypothesis}{nestedLogit}(model, ...)

\method{print}{dichotomies}(x, ...)

\method{predict}{nestedLogit}(object, newdata, model = c("nested", "dichotomies"), ...)

\method{print}{predictDichotomies}(x, ...)

\method{coef}{nestedLogit}(object, as.matrix = TRUE, ...)

\method{vcov}{nestedLogit}(object, as.matrix = FALSE, ...)

\method{anova}{nestedLogit}(object, object2, ...)

\method{print}{anova.nestedLogit}(x, ...)

\method{update}{nestedLogit}(object, formula, dichotomies, data, subset, contrasts, ...)

\method{as.matrix}{dichotomies}(x, ...)

\method{as.character}{dichotomies}(x, ...)

\method{as.matrix}{continuationDichotomies}(x, ...)

as.dichotomies(x, ...)

\method{as.dichotomies}{matrix}(x, ...)

\method{glance}{nestedLogit}(x, ...)

\method{tidy}{nestedLogit}(x, ...)
}
\arguments{
\item{x, object, object2, mod}{in most cases, an object of class \code{"nestedLogit"}.}

\item{\dots}{arguments to be passed down. In the case of \code{linearHypothesis},
the second argument is typically the \code{hypothesis.matrix}. See the
Details section of \code{\link[car]{linearHypothesis}}}

\item{model}{For the \code{predict} method, either \code{"nested"} (the default), in which case fitted probabilities
under the nested logit model are returned, or \code{"dichotomies"}, in which case
\code{\link{predict.glm}} is invoked for each binary logit model fit to the nested
dichotomies and a named list of the results is returned;
for \code{linearHypothesis}, an object of class \code{"nestedLogit"}.}

\item{newdata}{For the \code{predict} method, a data frame containing combinations of values of the predictors
at which fitted probabilities (or other quantities) are to be computed.}

\item{as.matrix}{if \code{TRUE} (the default for \code{coef}) return coefficients
as a matrix with one column for each nested dichotomy,
or coefficient covariances as a matrix with one row and column for each
combination of dichotomies and coefficients; if \code{FALSE} (the default for
\code{vcov}), return a list of coefficients or coefficient covariances
with one element for each dichotomy.}

\item{formula}{optional updated model formula.}

\item{dichotomies}{optional updated dichotomies object.}

\item{data}{optional updated data argument}

\item{subset}{optional updated subset argument.}

\item{contrasts}{optional updated contrasts argument.}
}
\description{
Various methods for processing \code{"nestedLogit"} and related objects.
Most of these are the standard methods for a model-fitting function.
\describe{
\item{\code{coef}, \code{vcov}}{Return the coefficients and their variance-covariance matrix respectively.}
\item{\code{Anova}}{Calculates type-II or type-III analysis-of-variance tables for \code{"nestedLogit"} objects.}
\item{\code{anova}}{Computes sequential analysis of variance (or deviance) tables for one or more fitted \code{"nestedLogit"}  objects.}
\item{\code{linearHypothesis}}{Computes Wald tests for linear hypotheses.}
\item{\code{update}}{Re-fit a \code{"nestedLogit"} model with a change in any of the \code{formula}, \code{dichotomies},
\code{data}, \code{subset}, or \code{contrasts}, arguments.}
\item{\code{predict}}{Computes predicted values from a fitted \code{"nestedLogit"} model.}
\item{\code{glance}}{Construct a single row summaries for the dichotomies \code{"nestedLogit"} model.}
\item{\code{tidy}}{Summarizes the terms in \code{"nestedLogit"} model.}
}
}
\details{
The \code{predict} method provides predicted values for two representations of the model.
\code{model = "nested"} gives the fitted probabilities for each of the response categories.
\code{model = "dichotomies"} gives the fitted log odds for each binary logit models in the
dichotomies.
}
\examples{
# define continuation dichotomies for level of education
cont.dichots <- continuationLogits(c("l.t.highschool",
                                     "highschool",
                                     "college",
                                     "graduate"))

# Show dichotomies in various forms
print(cont.dichots)
as.matrix(cont.dichots)
as.character(cont.dichots)

# fit a nested model for the GSS data examining education degree in relation to parent & year
data(GSS)
m <- nestedLogit(degree ~ parentdeg + year,
                 cont.dichots,
                 data=GSS)

coef(m)                             # coefficient estimates
sqrt(diag(vcov(m, as.matrix=TRUE))) # standard errors
print(m)
summary(m)

# broom methods
broom::glance(m)
broom::tidy(m)

# Anova tests
car::Anova(m) # type-II (partial) tests

# anova() and update() methods
anova(m) # type-I (sequential) tests
anova(update(m, . ~ . - year), m) # model comparison

# Wald test
car::linearHypothesis(m, c("parentdeghighschool", "parentdegcollege", 
                           "parentdeggraduate"))

# predicted probabilities and ploting
head(predict(m)) # fitted probabilities for first few cases
new <- expand.grid(parentdeg=c("l.t.highschool",  "highschool",
                               "college", "graduate"),
                   year=c(1972, 2016))
fit <- predict(m, newdata=new)
cbind(new, fit) # fitted probabilities at specific values of predictors
predictions <- predict(m, newdata=new, model="dichotomies", se.fit=TRUE)
predictions
predictions$above_l.t.highschool # on logit scale

}
\seealso{
\code{\link{nestedLogit}}, \code{\link{plot.nestedLogit}}
}
\author{
John Fox and Michael Friendly
}
\keyword{regression}
